+++ modules/apps/roles/roles-item-selector-api/src/main/java/com/liferay/roles/item/selector/OrganizationRoleItemSelectorCriterion.java
@@ -3,13 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.roles.item.selector.regular.role;
-
-import com.liferay.roles.item.selector.BaseRoleItemSelectorCriterion;
+package com.liferay.roles.item.selector;
 
 /**
  * @author Roberto Díaz
  */
-public class RegularRoleItemSelectorCriterion
+public class OrganizationRoleItemSelectorCriterion
 	extends BaseRoleItemSelectorCriterion {
 }
